feat(celery Wave 5 P1 chunk 1): §K.9 spec + aperag/indexing/llm.py relocate (legacy graphindex elimination first chunk)
Wave 5 task #26 first chunk per architect msg=10b5fae6 §K.9 spec
amendment + PM msg=af82797e dispatch (single Wave 5 PR with 5-phase
chunked-rotation commits, big-PR fast-landing per earayu2 msg=eced858d).

Two changes bundled (per PM msg=af82797e "Phase 1 first commit =
§K.9 spec paste + legacy graphindex elimination 实现起步"):

1. **§K.9 Wave 5 spec section** added to design pack:
   - 5-phase commit roadmap (16 acceptance items)
   - production-readiness 三类 layer per phase
   - architect direct ratify lane scope per phase
   - pre-check pattern lock per phase (per
     ``feedback_spec_lock_grep_verify_caller.md`` 双 pattern)
   - Layer 1 same-session continuation directive (per
     ``feedback_no_refresh_complete_all_tasks.md``)
   - Wave 5 acceptance summary (16 items + owner table)

2. **`aperag/indexing/llm.py` relocate** (§K.9.1 acceptance item 3):
   - new module owns canonical ``build_collection_llm_callable`` +
     ``render_extraction_prompt`` + ``ENTITY_RELATION_EXTRACTION``
     template + ``LLMCall`` type alias
   - legacy ``aperag/domains/knowledge_graph/graphindex/integration.py``
     and ``prompts.py`` re-export from new location during Wave 5
     deprecation window so legacy retrieval/curation callers keep
     working until Phase 1 close-out
   - ``aperag/indexing/graph_extractor.py`` updated to import from
     new ``aperag.indexing.llm`` module directly (instead of legacy
     graphindex bridge)
   - test ``test_graph_extractor.py`` monkey-patch sites updated to
     target ``aperag.indexing.llm`` module

Pre-check pattern 1 grep-verify caller cascade (per architect msg=
b26f64b2 chunk 4d Option C ruling, used as reference list for
Phase 1 migration scope):

- ``aperag/indexing/`` → 0 cross-references to legacy
  ``graphindex/storage/*`` (chunk 4d narrowed scope invariant
  maintained ✅)
- legacy ``graphindex/integration.py`` and ``prompts.py`` callers:
  ``retrieval/pipeline.py:85`` / ``knowledge_graph/service.py:69+`` /
  ``graph_curation/service.py:37`` / ``graph_curation/integration.py:21``
  / ``service/prompt_template_service.py:161`` —
  remaining migrations land in subsequent Phase 1 commits

Local gates green:
  - ``ruff check ./aperag ./tests`` clean
  - ``ruff format --check ./aperag ./tests`` clean (509 files)
  - ``pytest test_graph_extractor.py + test_full_indexing_pipeline.py
    + test_worker_factory.py + tests/unit_test/indexing/``: 179 passed
    / 2 skipped (Layer 2 stubs)

Production-readiness three-class layer (Phase 1 partial):
  - must-be-real: ``aperag.indexing.llm`` module is the canonical
    production home for the relocated LLM helpers (no import
    indirection through legacy package once retrieval/curation
    migration completes)
  - may-be-gated: legacy ``graphindex/{integration,prompts}.py``
    re-export shims kept active during Wave 5 deprecation window so
    cross-cutting refactor can land incrementally without breaking
    legacy retrieval/curation flow
  - fully-resolves: §K.9.1 Wave 5 acceptance item 3 (`aperag/indexing/llm.py`
    relocate); items 1 (legacy graphindex package elimination) and
    rest of Phase 1 cascade migration land in subsequent commits

Next Phase 1 commits: migrate ``retrieval/pipeline.py`` /
``knowledge_graph/service.py`` / ``graph_curation/*`` to §G.5 read
primitives; once all callers migrated → delete legacy
``graphindex/{storage, service, integration, engine, __init__}.py``
+ delete legacy tests; final commit verifies grep-zero invariant.
